Create MusicDownloader
This project is a Python-based music auto-searcher and downloader, utilizing the `youtube_dl` library. It simplifies the process of finding and downloading music by allowing users to input song names, which are then automatically retrieved from YouTube and saved to a specified directory. The program offers a straightforward command-line interface and enables users to customize the output path for downloaded music. It serves as a convenient tool for effortlessly accessing and enjoying a wide range of music content.